+++ src/db.js
@@ -492,6 +492,13 @@
             if (!serverEvents.includes(eventName)) {
                 throw new Error('Unrecognized event type ' + eventName);
             }
+            if (eventName === 'error') {
+                db.addEventListener(eventName, function (e) {
+                    e.preventDefault(); // Needed by Firefox to prevent hard abort with ConstraintError
+                    handler(e);
+                });
+                return;
+            }
             db.addEventListener(eventName, handler);
         };